feat(server): enforce naming policy on V2 create/update paths
Add NameValidator (server/util) as the shared enforcement point for
Constants.Name.PATTERN. V3NameValidator now delegates to it.

Apply NameValidator.validate() to all V2 create/update path variables:
- ServiceController: service name
- ServiceLabelController: service + label names, copy target
- ServiceAliasController: service + alias names
- StorageController: storage name

Read/delete paths are unaffected — existing legacy names remain accessible.
